Add new module spaces to contain PacketSpace and dependencies
This also helps reducing the size of the connection module.

Fold dedup module into it, since it is relatively small.
