--- quinn-proto/src/dedup.rs
@@ -1,131 +1,0 @@
-//! Packet deduplication
-
-use std::{cmp, mem};
-
-/// RFC4303-style sliding window packet number deduplicator.
-///
-/// A contiguous bitfield, where each bit corresponds to a packet number and the rightmost bit is
-/// always set. A set bit represents a packet that has been successfully authenticated. Bits left of
-/// the window are assumed to be set.
-///
-/// ```text
-/// ...xxxxxxxxx 1 0
-///     ^        ^ ^
-/// window highest next
-/// ```
-pub struct Dedup {
-    window: Window,
-    /// Lowest packet number higher than all yet authenticated.
-    next: u64,
-}
-
-/// Inner bitfield type.
-///
-/// Because QUIC never reuses packet numbers, this only needs to be large enough to deal with
-/// packets that are reordered but still delivered in a timely manner.
-type Window = u128;
-
-/// Number of packets tracked by `Dedup`.
-const WINDOW_SIZE: u64 = 1 + mem::size_of::<Window>() as u64 * 8;
-
-impl Dedup {
-    /// Construct an empty window positioned at the start.
-    pub fn new() -> Self {
-        Self { window: 0, next: 0 }
-    }
-
-    /// Highest packet number authenticated.
-    fn highest(&self) -> u64 {
-        self.next - 1
-    }
-
-    /// Record a newly authenticated packet number.
-    ///
-    /// Returns whether the packet might be a duplicate.
-    pub fn insert(&mut self, packet: u64) -> bool {
-        if let Some(diff) = packet.checked_sub(self.next) {
-            // Right of window
-            self.window = (self.window << 1 | 1)
-                .checked_shl(cmp::min(diff, u64::from(u32::max_value())) as u32)
-                .unwrap_or(0);
-            self.next = packet + 1;
-            false
-        } else if self.highest() - packet < WINDOW_SIZE {
-            // Within window
-            if let Some(bit) = (self.highest() - packet).checked_sub(1) {
-                // < highest
-                let mask = 1 << bit;
-                let duplicate = self.window & mask != 0;
-                self.window |= mask;
-                duplicate
-            } else {
-                // == highest
-                true
-            }
-        } else {
-            // Left of window
-            true
-        }
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-
-    #[test]
-    fn sanity() {
-        let mut dedup = Dedup::new();
-        assert!(!dedup.insert(0));
-        assert_eq!(dedup.next, 1);
-        assert_eq!(dedup.window, 0b1);
-        assert!(dedup.insert(0));
-        assert_eq!(dedup.next, 1);
-        assert_eq!(dedup.window, 0b1);
-        assert!(!dedup.insert(1));
-        assert_eq!(dedup.next, 2);
-        assert_eq!(dedup.window, 0b11);
-        assert!(!dedup.insert(2));
-        assert_eq!(dedup.next, 3);
-        assert_eq!(dedup.window, 0b111);
-        assert!(!dedup.insert(4));
-        assert_eq!(dedup.next, 5);
-        assert_eq!(dedup.window, 0b11110);
-        assert!(!dedup.insert(7));
-        assert_eq!(dedup.next, 8);
-        assert_eq!(dedup.window, 0b11110100);
-        assert!(dedup.insert(4));
-        assert!(!dedup.insert(3));
-        assert_eq!(dedup.next, 8);
-        assert_eq!(dedup.window, 0b11111100);
-        assert!(!dedup.insert(6));
-        assert_eq!(dedup.next, 8);
-        assert_eq!(dedup.window, 0b11111101);
-        assert!(!dedup.insert(5));
-        assert_eq!(dedup.next, 8);
-        assert_eq!(dedup.window, 0b11111111);
-    }
-
-    #[test]
-    fn happypath() {
-        let mut dedup = Dedup::new();
-        for i in 0..(2 * WINDOW_SIZE) {
-            assert!(!dedup.insert(i));
-            for j in 0..=i {
-                assert!(dedup.insert(j));
-            }
-        }
-    }
-
-    #[test]
-    fn jump() {
-        let mut dedup = Dedup::new();
-        dedup.insert(2 * WINDOW_SIZE);
-        assert!(dedup.insert(WINDOW_SIZE));
-        assert_eq!(dedup.next, 2 * WINDOW_SIZE + 1);
-        assert_eq!(dedup.window, 0);
-        assert!(!dedup.insert(WINDOW_SIZE + 1));
-        assert_eq!(dedup.next, 2 * WINDOW_SIZE + 1);
-        assert_eq!(dedup.window, 1 << (WINDOW_SIZE - 2));
-    }
-}
